neatogen stress_majorization_with_hierarchy: fix memory leak of 'f_storage'
CodeChecker says:

  [MEDIUM] lib/neatogen/constrained_majorization.c:279:30: Potential leak of
    memory pointed to by 'f_storage' [unix.Malloc]
      constant_term = (float) (n * (n - 1) / 2);
                               ^

When examining the surrounding code, it first appears this is a false positive
as the memory stored in `f_storage` is freed through `coords[0]`. But I think
what CodeChecker is pointing out is that when `dim == 0` a compliant allocator
could return non-null for the `f_storage` allocation but null for the `coords`
allocation. This would result in leaking the `f_storage` allocation. At time of
writing, I am not aware of any allocators with this behavior, though technically
I think this would be ISO C17 compliant.

Reported-by: CodeChecker 6.26.2