quartz plugin: abbreviate 'memory_data_provider_callbacks' initialization
At time of writing, the layout of `CGDataProviderDirectCallbacks` seems
unspecified.¹ It seems unlikely this would change in future, as it would break
any pre-compiled software like this code. However, note that the members of
`CGDataProviderDirectCallbacks` are currently documented in a different order
than their layout. This change makes the source assume less about the layout of
this structure.

¹ https://developer.apple.com/documentation/coregraphics/cgdataproviderdirectcallbacks?language=objc