core plugin: remove unused 'render_features_json1'
008ea64e1cc3d1d349b62145bb7ddb33d29625ba seems to have made a typo when adding
this, making the new format `json0` simply a copy of the existing `json`. This
persisted through 2b43a2246e9c4249c80f630077c511d6470dac83 which renamed this,
still seemingly oblivious to it being unused.

There have been no reports about the (mis)behavior of this `json0` format in the
meantime and no one seems to have noticed it is identical to `json`. So lets
just remove this loose end and leave the behavior as-is. At the next major
version bump, we should remove the `json0` alias altogether.