osage layout: abbreviate assignments to 'rootbb'
As odd as it looks, the last assignment really did (and does) seem to be setting
the lower left corner of the bounding box to 0. This seems to have been the case
ever since the first introduction of this code, so we may as well do this in a
clearer way.