gen_version: fix: pass 'CHANGELOG.md' to Git as a relative path
When running under MinGW, the path to CHANGELOG.md is resolved as a Windows path
but then passed onwards to the MinGW `git` binary which expects Unix paths. A
side effect of this is that the MinGW jobs in CI believed they were asked for
version control history about a file not under version control:

  fatal: 'C:/GitLab-Runner/builds/smattr/graphviz/CHANGELOG.md' is outside
    repository at '/c/GitLab-Runner/builds/smattr/graphviz'

Though this is listed as “fatal” it does not actually terminate CI because
gen_version.py attempts to handle this case as if Git were missing and/or
Graphviz is being built from a release tarball.

This change uses a relative path that both the MinGW `git` and the Windows `git`
(should we use this in future) understand. Note that this still leaves the foot
gun that Git failures here can go unnoticed. In future this should probably be
improved to hard-fail in CI where we never expect Git to be missing or to be
building from a non-Git-checkout.

Gitlab: fixes #2789