Autotools: fix: do not build PHP bindings when 'php-config' is not found
When `php` was available, building of the PHP bindings would be enabled,
unconditional on whether `php-config` had been found. The result of this would
be that compilation failed, as none of the paths needed to build the PHP
bindings had been detected.

This appears to have been a mistake in 4129afb8db3039551f85a4e38fd5cee623e5d729.

This change leaves `php` discovery in place because this is used in a test
target. The test target is not currently run but may be resurrected in future.

Build system support for the PHP bindings has had a rocky history. It is
suspected this is not the last bug that will be discovered in this logic.

Gitlab: fixes #2788
Reported-by: FeRD (Frank Dana) <ferdnyc@gmail.com>