sfio sfvscanf: remove use of 'n' as a loop temporary
This was most likely an optimization to avoid multiple memory reads when
dereferencing `form`. This is no longer a concern on contemporary compilers that
are able to de-dupe these reads themselves.

This is a step towards reducing the scope of `n`, making it more likely
compilers are able to catch issues like those addressed in the previous commits.