sfio sfvscanf: remove another dead store to 'n'
CodeChecker says:

  [LOW] lib/sfio/sfvscanf.c:240:3: Value stored to 'n' is never read
    [deadcode.DeadStores]
      n = ++argn;
      ^

It seems this became unused in 1d3e1bd053724cdf3cf331acb86d1f5a5c7e6825, but the
prolific reuse of this variable made it difficult for humans and compilers to
spot this.

Note that this change also makes `argn` unused, so we can remove it too.

Reported-by: CodeChecker 6.26.2