osage layout: remove manual memory management of 'gs'
Given the allocation of this array was incorrect in the past (see
e5a5360ca2af8eb21338eeb3cdc38ba408d8a6d9), it seems a good idea to remove
assumptions about the maximum size of this array from this code.

During testing this change, it was discovered that the “reposition children”
loop was actually processing unset (but zeroed) boxes beyond the end of the list
in some circumstances. Hence why this change also changes the bounds check of
this loop.