CI: fix processing of disabled config lines containing ':'
The upcoming addition of a Ubuntu 25.10 Autotools job produces config output
that looks like:

  …
  smyrna:        No (requires: gtk+ gtkglext glade)
  …

This logic was splitting on ":", expecting there to be 1 or 2 chunks as a
result. However this line generates 3 due to the ":" in the explanation. This
change works around the situation by discarding anything after the second chunk.