redhat: fix: conditionally run config setup for graphviz-plugins-x
Similar to the situation described in the previous commit, graphviz-plugins-x
does not depend on graphviz-plugins-nox, the package that installs `dot`. It
seems to coincidentally find `dot` available because the top level package
depends on graphviz-plugins-nox and graphviz-plugins-x in that order. However I
can find no guarantee `Requires` lines need to be processed in order.

This change ensures a consistent plugin discovery outcome regardless of which
order graphviz-plugins-nox and graphviz-plugins-x are installed in.