redhat: fix: conditionally run config setup for non-dot depending plugins
A number of the plugin packages are defined without depending on
graphviz-plugins-nox, the package that installs `dot`. I do not know why you
would want to install a plugin without a way to use it, but nevertheless this
seems historically possible. A side effect of this lack of dependencies on `dot`
being installed is that the post-run action of these packages can run without
`dot` available. The ordering seems to have coincidentally worked out in the
past but Fedora 43 installs the packages in a different order resulting in
failure:

  D: %post(graphviz-plugins-gd-14.1.2~dev.20251230.0728-1.fc43.x86_64): scriptlet start
  D: %post(graphviz-plugins-gd-14.1.2~dev.20251230.0728-1.fc43.x86_64): execv(/bin/sh) pid 32
  + LD_LIBRARY_PATH=/usr/lib64
  + /usr/bin/dot -c
  /var/tmp/rpm-tmp.k2tpkb: line 2: /usr/bin/dot: No such file or directory
  D: %post(graphviz-plugins-gd-14.1.2~dev.20251230.0728-1.fc43.x86_64): waitpid(32) rc 32 status 7f00
  warning: %post(graphviz-plugins-gd-14.1.2~dev.20251230.0728-1.fc43.x86_64) scriptlet failed, exit status 127

This change fixes this situation for plugins-gd, plugins-kitty, plugins-vt,
plugins-webp, and plugins-devil